+++ docs/src/changelog.md
@@ -16,6 +16,7 @@ Changelog](https://keepachangelog.com).
 - A new `Gssapi` module to help with [GSSAPI support](@ref). In particular,
   [`Gssapi.principal_name()`](@ref) was added to get the name of the default
   principal if one is available ([#6]).
+- An experimental [`authenticate()`](@ref) function to simplify authenticating ([#7]).
 
 ### Changed
 
@@ -23,6 +24,8 @@ Changelog](https://keepachangelog.com).
   they got a `AuthStatus_Error` from libssh. This can be disabled by passing
   `throw_on_error=false` ([#6]).
 - `gssapi_available()` was renamed to [`Gssapi.isavailable()`](@ref) ([#6]).
+- [`userauth_kbdint_getprompts()`](@ref) returns a vector of `KbdintPrompt`
+  objects instead of tuples ([#7]).
 
 ### Fixed